Move error reporter to errors folder
