+++ src/components/Cluster/ClusterDetail/NodePool.js
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import * as nodePoolActions from 'actions/nodePoolActions';
 import { spinner } from 'images';
-import ErrorReporter from 'lib/ErrorReporter';
+import ErrorReporter from 'lib/errors/ErrorReporter';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
