+++ src/lib/errors/CustomError.js
@@ -1,4 +1,4 @@
-import ErrorReporter from 'lib/ErrorReporter';
+import ErrorReporter from 'lib/errors/ErrorReporter';
 import StackTraceGPS from 'stacktrace-gps';
 import StackTrace from 'stacktrace-js';
