+++ src/lib/errors/ErrorReporter.js
